+++ src/api/boot/mod.rs
@@ -118,30 +118,3 @@ impl AsyncRobot {
         Ok(self.go(get_config(server_number)).await?.0)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    #[cfg(feature = "non-disruptive-tests")]
-    mod non_disruptive_tests {
-        use serial_test::serial;
-        use tracing::info;
-        use tracing_test::traced_test;
-
-        #[tokio::test]
-        #[traced_test]
-        #[serial(boot_configuration)]
-        async fn test_get_boot_configuration() {
-            let _ = dotenvy::dotenv().ok();
-
-            let robot = crate::AsyncRobot::default();
-
-            let servers = robot.list_servers().await.unwrap();
-            info!("{servers:#?}");
-
-            if let Some(server) = servers.first() {
-                let config = robot.get_boot_config(server.id).await.unwrap();
-                info!("{config:#?}");
-            }
-        }
-    }
-}